be explicit about type in a redim
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6000 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ custom.bas
@@ -83,7 +83,7 @@ REDIM gen(360)
 DIM gen_reld_doc as DocPtr
 REDIM buffer(16384)
 REDIM master(255) as RGBcolor
-REDIM uilook(uiColorLast)
+REDIM uilook(uiColorLast) as integer
 REDIM boxlook(uiBoxLast) as BoxStyle
 DIM statnames() as string
 REDIM herotags(59) as HeroTagsCache
